Fix starred story parsing using incorrect JSON field
